move dealing with tool results to tool_utils, move restart utils to util from restart tool
