+++ agent/tools/restart_program_tool.py
@@ -1,9 +1,6 @@
 # restart_program_tool.py
 
 import json
-import os
-import pickle
-import sys
 
 from agent.tools.base_tool import ToolDefinition
 
@@ -43,47 +40,6 @@ def restart_program(input_data: dict) -> str:
         "agent_initiated": True  # Flag to indicate this restart was initiated by the agent
     }
     return json.dumps(result)
-
-
-def load_conversation_context(save_file: str = "conversation_context.pkl"):
-    """
-    Call this once at your program’s entrypoint to restore a prior session.
-    """
-    if os.path.exists(save_file):
-        with open(save_file, "rb") as f:
-            context = pickle.load(f)
-        return context
-    return None
-
-
-def save_conversation(conversation, save_file: str):
-    """Save the conversation context to a file"""
-    try:
-        with open(save_file, 'wb') as f:
-            pickle.dump(conversation, f)
-        return True
-    except Exception as e:
-        error_message = f"Error saving conversation: {str(e)}"
-        print(error_message)
-        try:
-            with open("error.txt", "a") as f:
-                import datetime
-                timestamp = datetime.datetime.now().strftime("%Y-%m-%d %H:%M:%S")
-                f.write(f"\n[{timestamp}] ERROR: {error_message}\n")
-        except Exception:
-            pass  # Silently fail if we can't log to error.txt
-        return False
-
-
-def save_conv_and_restart(conversation, save_file: str = "conversation_context.pkl"):
-    save_conversation(conversation, save_file)
-
-    # Set a flag to indicate we're intentionally restarting
-    sys.is_restarting = True
-    
-    # re-exec in-place:
-    python = sys.executable
-    os.execv(python, [python] + sys.argv)
 
 
 # ------------------------------------------------------------------
